+++ IntegrationTests/Helpers/Helpers.swift
@@ -1,16 +1,10 @@
 import Foundation
 
-@objc class Helpers : NSObject {
+class Helpers: NSObject {
 
-    private static let _sharedInstance = Helpers()
+    static let shared = Helpers()
 
-    private override init() {}
-
-    @objc class func sharedInstance() -> Helpers {
-        return Helpers._sharedInstance
-    }
-
-    @objc func futureYear() -> String {
+    func futureYear() -> String {
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "yyyy"
         return dateFormatter.string(from: Date())