+++ UnitTests/BraintreeThreeDSecureTests/BTConfiguration+ThreeDSecure_Tests.swift
@@ -1,5 +1,5 @@
-import Foundation
 import XCTest
+@testable import BraintreeCore
 @testable import BraintreeThreeDSecure
 
 class BTConfiguration_ThreeDSecure_Tests: XCTestCase {